backend/requirements.txt
fastapi
uvicorn
openpyxl
passlib[bcrypt]
python-jose[cryptography]
filelock
bcrypt<4
